+++ src/edu/hitsz/aircraft/AbstractEnemyAircraft.java
@@ -1,7 +1,7 @@
 package edu.hitsz.aircraft;
 
 public abstract class AbstractEnemyAircraft extends AbstractAircraft{
-    public AbstractEnemyAircraft(int locationX, int locationY, int speedX, int speedY, int hp){
+    public AbstractEnemyAircraft(int locationX, int locationY, int speedX, double speedY, int hp){
         super(locationX, locationY, speedX, speedY, hp);
     }
 }
