+++ src/edu/hitsz/aircraft/BossEnemyFactory.java
@@ -4,7 +4,7 @@ import edu.hitsz.strategy.ScatteredShoot;
 
 public class BossEnemyFactory implements EnemyFactory{
     @Override
-    public AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, int speedY, int hp){
+    public AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, double speedY, int hp){
         return new BossEnemy(locationX,locationY,speedX,speedY,hp,new ScatteredShoot());
     }
 }
