+++ src/edu/hitsz/aircraft/EliteEnemyFactory.java
@@ -4,7 +4,7 @@ import edu.hitsz.strategy.DirectShoot;
 
 public class EliteEnemyFactory implements EnemyFactory{
     @Override
-    public AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, int speedY, int hp){
+    public AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, double speedY, int hp){
         return new EliteEnemy(locationX,locationY,speedX,speedY,hp,new DirectShoot());
     }
 }
