+++ src/edu/hitsz/aircraft/EnemyFactory.java
@@ -3,5 +3,5 @@ package edu.hitsz.aircraft;
 
 
 public interface EnemyFactory {
-     AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, int speedY, int hp);
+     AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, double speedY, int hp);
 }
