+++ src/edu/hitsz/aircraft/MobEnemyFactory.java
@@ -2,7 +2,7 @@ package edu.hitsz.aircraft;
 
 public class MobEnemyFactory implements EnemyFactory{
     @Override
-    public AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, int speedY, int hp){
+    public AbstractEnemyAircraft createEnemy(int locationX, int locationY, int speedX, double speedY, int hp){
         return new MobEnemy(locationX,locationY,speedX,speedY,hp);
     }
 }
